pr7 refactor queue and accessories
